backend/order_service/requirements.txt
asgiref==3.11.0
Django==4.2.27
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.2.13
PyJWT==2.10.1
sqlparse==0.5.4
typing_extensions==4.15.0
kombu
pika
